provide code and reason values to CloseEvents
Add support for differentiating between different types of WebSocket closures by passing `close` and `reason` values to `ws.close()`. This is helpful, because a client may wish to perform different reconneciton logic depending on the reason for disconnection.

In particular, it is useful to be able to differentiate intentional disconnections where the server wishes to indicate to the client that it should not reconnect at all or the client should delay for a more extended time before it reconnects. If, for example, a client's application version is no longer supported, the server may wish to disconnect with a reason that indicates the client should not attempt reconnection and it should refresh or update its code first.